Preliminary Data Preparation

Preliminary Data Preparation

Usage

DataOU(
  data,
  observed,
  id,
  time,
  insert_na = FALSE,
  initial_na = TRUE,
  center = FALSE,
  scale = FALSE
)

Arguments

data

Data frame. A data frame object of data for potentially multiple subjects that contain a column of subject ID numbers (i.e., an ID variable), a column indicating subject-specific measurement occasions (i.e., a TIME variable), at least one column of observed values.

observed

Character vector. A vector of character strings of the names of the observed variables in the data.

id

Character string. A character string of the name of the ID variable in the data.

time

Character string. A character string of the name of the TIME variable in the data.

insert_na

Logical. Insert NA to observed variables for existing time points.

initial_na

Logical. Iteratively remove rows where any observed variable for the first time point has NA.

center

Logical. If center = TRUE, mean center by id.

scale

Logical. If scale = TRUE, standardize by id.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

data <- DataOU(
  data = bivariate_ou,
  observed = c("y1", "y2"),
  id = "id",
  time = "time"
)
summary(data)
#>        id             time             y1                 y2          
#>  Min.   : 1.00   Min.   : 0.00   Min.   :-9.38679   Min.   :-15.6376  
#>  1st Qu.: 5.75   1st Qu.:12.47   1st Qu.: 0.05093   1st Qu.:  0.6164  
#>  Median :10.50   Median :24.95   Median : 3.08874   Median :  3.5348  
#>  Mean   :10.50   Mean   :24.95   Mean   : 2.97662   Mean   :  3.3015  
#>  3rd Qu.:15.25   3rd Qu.:37.42   3rd Qu.: 5.85255   3rd Qu.:  6.2228  
#>  Max.   :20.00   Max.   :49.90   Max.   :16.11338   Max.   : 16.6516