Fit the Ornstein–Uhlenbeck Model

This is a wrapper function that makes fitting the Ornstein–Uhlenbeck model convenient using the dynr package.

Usage

FitOU(
  data,
  observed,
  id,
  time,
  mu0 = NULL,
  sigma0 = NULL,
  mu_start = NULL,
  phi_start = NULL,
  sigma_start = NULL,
  theta_start = NULL,
  ...
)

Arguments

data

Data frame. A data frame object of data for potentially multiple subjects that contain a column of subject ID numbers (i.e., an ID variable), a column indicating subject-specific measurement occasions (i.e., a TIME variable), at least one column of observed values.

observed

Character vector. A vector of character strings of the names of the observed variables in the data.

id

Character string. A character string of the name of the ID variable in the data.

time

Character string. A character string of the name of the TIME variable in the data.

mu0

Numeric vector. Mean of initial latent variable values (\(\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}\)). If mu0 = NULL, a vector of zeros is used.

sigma0

Numeric matrix. Covariance matrix of initial latent variable values (\(\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}\)). If sigma0 = NULL, an identity matrix is used.

mu_start

Numeric vector. Starting values of the mu vector, that is, the long-term mean or equilibrium level. If mu_start = NULL, a vector of zeros is used.

phi_start

Numeric matrx. Starting values of the phi matrix, that is, the rate of mean reversion, determining how quickly the variable returns to its mean. If phi_start = NULL, a matrix of zeros is used.

sigma_start

Numeric matrx. Starting values of the sigma matrix, that is, the matrix of volatility or randomness in the process. If sigma_start = NULL, an identity matrix is used.

theta_start

Numeric matrx. Starting values of the theta matrix, that is, the measurement error covariance matrix (\(\boldsymbol{\Theta}\)). If theta_start = NULL, an identity matrix is used.

...

Additional arguments to pass to dynr::dynr.cook().

Details

The measurement model is given by $$ \mathbf{y}_{i, t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{i, t} + \boldsymbol{\varepsilon}_{i, t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) $$

where \(\mathbf{y}_{i, t}\), \(\boldsymbol{\eta}_{i, t}\), and \(\boldsymbol{\varepsilon}_{i, t}\) are random variables and \(\boldsymbol{\nu}\), \(\boldsymbol{\Lambda}\), and \(\boldsymbol{\Theta}\) are model parameters. \(\mathbf{y}_{i, t}\) is a vector of observed random variables at time \(t\) and individual \(i\), \(\boldsymbol{\eta}_{i, t}\) is a vector of latent random variables at time \(t\) and individual \(i\), and \(\boldsymbol{\varepsilon}_{i, t}\) is a vector of random measurement errors at time \(t\) and individual \(i\), while \(\boldsymbol{\nu}\) is a vector of intercept, \(\boldsymbol{\Lambda}\) is a matrix of factor loadings, and \(\boldsymbol{\Theta}\) is the covariance matrix of \(\boldsymbol{\varepsilon}\).

The dynamic structure is given by

$$ \mathrm{d} \boldsymbol{\eta}_{i, t} = \boldsymbol{\Phi} \left( \boldsymbol{\mu} - \boldsymbol{\eta}_{i, t} \right) \mathrm{d}t + \boldsymbol{\Sigma}^{\frac{1}{2}} \mathrm{d} \mathbf{W}_{i, t} $$

where \(\boldsymbol{\mu}\) is the long-term mean or equilibrium level, \(\boldsymbol{\Phi}\) is the rate of mean reversion, determining how quickly the variable returns to its mean, \(\boldsymbol{\Sigma}\) is the matrix of volatility or randomness in the process, and \(\mathrm{d}\boldsymbol{W}\) is a Wiener process or Brownian motion, which represents random fluctuations.

References

Chow, S.-M., Losardo, D., Park, J., & Molenaar, P. C. M. (2023). Continuous-time dynamic models: Connections to structural equation models and other discrete-time models. In R. H. Hoyle (Ed.), Handbook of structural equation modeling (2nd ed.). The Guilford Press.

Ou, L., Hunter, M. D., & Chow, S.-M. (2019). What's for dynr: A package for linear and nonlinear dynamic modeling in R. The R Journal, 11(1), 91. doi:10.32614/rj-2019-012

Uhlenbeck, G. E., & Ornstein, L. S. (1930). On the theory of the brownian motion. Physical Review, 36(5), 823–841. doi:doi.org/10.1103/physrev.36.823

Author

Ivan Jacob Agaloos Pesigan

Examples

if (FALSE) {
FitOU(
  data = bivariate_ou,
  observed = c("y1", "y2"),
  id = "id",
  time = "time",
  verbose = FALSE
)
}